openai-chat/requirements.txt
httpx
markdown
python-dotenv
requests